pyproject.toml
[project]
name = "data-engineering-course"
version = "0.1.0"
description = "This repository has the necessary files for a data engineering course offered by @tetelio"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.57",
    "matplotlib>=3.10.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
